+++ app/frontend/dashboard.html
@@ -28,31 +28,31 @@
             
             <div class="sidebar-content">
                 <nav class="sidebar-menu">
-                    <a href="dashboard.html" class="sidebar-menu-item active">
+                    <a href="/dashboard" class="sidebar-menu-item active">
                         <span class="sidebar-menu-icon">
                             <i class="fas fa-chart-line"></i>
                         </span>
                         <span class="sidebar-menu-text">Dashboard</span>
                     </a>
-                    <a href="capture.html" class="sidebar-menu-item">
+                    <a href="/capture" class="sidebar-menu-item">
                         <span class="sidebar-menu-icon">
                             <i class="fas fa-camera"></i>
                         </span>
                         <span class="sidebar-menu-text">Capture</span>
                     </a>
-                    <a href="handbook.html" class="sidebar-menu-item">
+                    <a href="/handbook" class="sidebar-menu-item">
                         <span class="sidebar-menu-icon">
                             <i class="fas fa-book"></i>
                         </span>
                         <span class="sidebar-menu-text">Handbook</span>
                     </a>
-                    <a href="account.html" class="sidebar-menu-item">
+                    <a href="/account" class="sidebar-menu-item">
                         <span class="sidebar-menu-icon">
                             <i class="fas fa-user"></i>
                         </span>
                         <span class="sidebar-menu-text">Account</span>
                     </a>
-                    <a href="login.html" class="sidebar-menu-item">
+                    <a href="/login" class="sidebar-menu-item">
                         <span class="sidebar-menu-icon">
                             <i class="fas fa-sign-out-alt"></i>
                         </span>
@@ -367,7 +367,7 @@
                         <h2 class="section-title">Common Diseases</h2>
                         <p class="section-subtitle">Most frequent plant issues in your garden</p>
                     </div>
-                    <a href="handbook.html" class="btn btn-primary">View All Diseases</a>
+                    <a href="/handbook" class="btn btn-primary">View All Diseases</a>
                 </div>
                 
                 <div class="diseases-grid">
