+++ app/frontend/index.html
@@ -16,7 +16,7 @@
         <div class="container">
             <nav class="navbar">
                 <div class="logo">
-                    <a href="index.html">
+                    <a href="/">
                         <img src="assets/logo/green-logo.png" alt="DAHON Logo">
                         <span>DAHON</span>
                     </a>
@@ -67,7 +67,7 @@
                     </div>
                     <div class="hero-buttons">
                         <a href="signup.html" class="btn btn-primary">Get Started <i class="fas fa-arrow-right"></i></a>
-                        <a href="login.html" class="btn btn-secondary">Learn More</a>
+                        <a href="/login" class="btn btn-secondary">Learn More</a>
                     </div>
                     <div class="hero-badge">
                         <div class="badge-content">
@@ -226,7 +226,7 @@
 </div>
             
             <div class="categories-footer">
-                <a href="login.html" class="btn btn-primary view-all-btn">
+                <a href="/login" class="btn btn-primary view-all-btn">
                     <span>View All Plant Diseases</span>
                     <i class="fas fa-arrow-right"></i>
                 </a>
