+++ app/frontend/plant-details.html
@@ -26,31 +26,31 @@
             
             <div class="sidebar-content">
                 <nav class="sidebar-menu">
-                    <a href="dashboard.html" class="sidebar-menu-item">
+                    <a href="/dashboard" class="sidebar-menu-item">
                         <span class="sidebar-menu-icon">
                             <i class="fas fa-chart-line"></i>
                         </span>
                         <span class="sidebar-menu-text">Dashboard</span>
                     </a>
-                    <a href="capture.html" class="sidebar-menu-item">
+                    <a href="/capture" class="sidebar-menu-item">
                         <span class="sidebar-menu-icon">
                             <i class="fas fa-camera"></i>
                         </span>
                         <span class="sidebar-menu-text">Capture</span>
                     </a>
-                    <a href="handbook.html" class="sidebar-menu-item">
+                    <a href="/handbook" class="sidebar-menu-item">
                         <span class="sidebar-menu-icon">
                             <i class="fas fa-book"></i>
                         </span>
                         <span class="sidebar-menu-text">Handbook</span>
                     </a>
-                    <a href="account.html" class="sidebar-menu-item">
+                    <a href="/account" class="sidebar-menu-item">
                         <span class="sidebar-menu-icon">
                             <i class="fas fa-user"></i>
                         </span>
                         <span class="sidebar-menu-text">Account</span>
                     </a>
-                    <a href="login.html" class="sidebar-menu-item">
+                    <a href="/login" class="sidebar-menu-item">
                         <span class="sidebar-menu-icon">
                             <i class="fas fa-sign-out-alt"></i>
                         </span>
