+++ app/frontend/signup.html
@@ -22,21 +22,21 @@
         <!-- Header -->
         <header class="site-header">
             <div class="container">
-                <a href="index.html" class="logo">
+                <a href="/" class="logo">
                     <img src="assets/logo/green-logo.png" alt="DAHON Logo" class="logo-icon">
                     <span>DAHON</span>
                 </a>
                 <nav class="main-nav">
                     <ul>
-                        <li><a href="index.html">Home</a></li>
-                        <li><a href="index.html">Categories</a></li>
-                        <li><a href="index.html">About us</a></li>
-                        <li><a href="index.html">FAQ</a></li>
+                        <li><a href="/">Home</a></li>
+                        <li><a href="/">Categories</a></li>
+                        <li><a href="/">About us</a></li>
+                        <li><a href="/">FAQ</a></li>
                     </ul>
                 </nav>
                 <div class="auth-links">
-                    <a href="login.html" class="btn btn-text">Log In</a>
-                    <a href="signup.html" class="btn btn-primary active">Sign up</a>
+                    <a href="/login" class="btn btn-text">Log In</a>
+                    <a href="/register" class="btn btn-primary active">Sign up</a>
                 </div>
                 <button class="menu-toggle" aria-label="Toggle menu" aria-expanded="false">
                     <span></span>
@@ -238,7 +238,7 @@
                         </div>
                         
                         <div class="auth-footer">
-                            <p>Already have an account? <a href="login.html">Log in</a></p>
+                            <p>Already have an account? <a href="/login">Log in</a></p>
                         </div>
                     </div>
                 </div>
